adjusted service interfaces and contracts. added parameter checks.

+++ src/blocks/blockService.ts
@@ -1,4 +1,5 @@
-﻿import * as Utils from "../utils";
+﻿import { Bag } from "./../bag";
+import * as Utils from "../utils";
 import { IObjectStorage } from "../persistence";
 import { IBlockService } from "./IBlockService";
 import { Contract } from "../contract";
@@ -7,13 +8,15 @@ import { BlockContract } from "./blockContract";
 const blockPath = "blocks";
 
 export class BlockService implements IBlockService {
-    private readonly objectStorage: IObjectStorage;
-
-    constructor(objectStorage: IObjectStorage) {
-        this.objectStorage = objectStorage;
-    }
+    constructor(
+        private readonly objectStorage: IObjectStorage
+    ) { }
 
     public getBlockByKey(key: string): Promise<BlockContract> {
+        if (!key) {
+            throw new Error(`Parameter "key" not specified.`);
+        }
+
         if (!key.startsWith(blockPath)) {
             return null;
         }
@@ -21,17 +24,16 @@ export class BlockService implements IBlockService {
     }
 
     public async search(pattern: string): Promise<BlockContract[]> {
-        return await this.objectStorage.searchObjects<BlockContract>(blockPath);
+        const result = await this.objectStorage.searchObjects<Bag<BlockContract>>(blockPath, ["title"], pattern);
+        return Object.keys(result).map(key => result[key]);
     }
 
     public async deleteBlock(block: BlockContract): Promise<void> {
-        try {
-            await this.objectStorage.deleteObject(block.key);
+        if (!block) {
+            throw new Error(`Parameter "block" not specified.`);
         }
-        catch (error) {
-            // TODO: Do proper handling.
-            console.warn(error);
-        }
+
+        await this.objectStorage.deleteObject(block.key);
     }
 
     public async createBlock(title: string, description: string, content: Contract): Promise<void> {
@@ -42,12 +44,16 @@ export class BlockService implements IBlockService {
             title: title,
             description: description,
             content: content
-        }
+        };
 
         await this.objectStorage.updateObject(key, block);
     }
 
     public updateBlock(block: BlockContract): Promise<void> {
+        if (!block) {
+            throw new Error(`Parameter "block" not specified.`);
+        }
+
         return this.objectStorage.updateObject(block.key, block);
     }
 }
