+++ src/pages/pageService.ts
@@ -1,4 +1,5 @@
-﻿import * as Utils from "../utils";
+﻿import { Bag } from "./../bag";
+import * as Utils from "../utils";
 import { PageContract, IPageService } from "../pages";
 import { IObjectStorage } from "../persistence";
 import { IBlockService } from "../blocks";
@@ -14,24 +15,39 @@ export class PageService implements IPageService {
         private readonly blockService: IBlockService
     ) { }
 
-    private async searchByTags(tags: string[], tagValue: string, startAtSearch: boolean): Promise<PageContract[]> {
-        return await this.objectStorage.searchObjects<PageContract>(pagesPath, tags, tagValue, startAtSearch);
+    private async searchByProperties(properties: string[], value: string): Promise<PageContract[]> {
+        const result = await this.objectStorage.searchObjects<Bag<PageContract>>(pagesPath, properties, value);
+        return Object.keys(result).map(key => result[key]);
     }
 
     public async getPageByPermalink(permalink: string): Promise<PageContract> {
-        const pages = await this.objectStorage.searchObjects<any>(pagesPath, ["permalink"], permalink);
+        if (!permalink) {
+            throw new Error(`Parameter "permalink" not specified.`);
+        }
+
+        const result = await this.objectStorage.searchObjects<Bag<PageContract>>(pagesPath, ["permalink"], permalink);
+        const pages = Object.keys(result).map(key => result[key]);
+
         return pages.length > 0 ? pages[0] : null;
     }
 
     public async getPageByKey(key: string): Promise<PageContract> {
+        if (!key) {
+            throw new Error(`Parameter "key" not specified.`);
+        }
+
         return await this.objectStorage.getObject<PageContract>(key);
     }
 
     public search(pattern: string): Promise<PageContract[]> {
-        return this.searchByTags(["title"], pattern, true);
+        return this.searchByProperties(["title"], pattern);
     }
 
     public async deletePage(page: PageContract): Promise<void> {
+        if (!page) {
+            throw new Error(`Parameter "page" not specified.`);
+        }
+
         const deleteContentPromise = this.objectStorage.deleteObject(page.contentKey);
         const deletePagePromise = this.objectStorage.deleteObject(page.key);
 
@@ -62,16 +78,32 @@ export class PageService implements IPageService {
     }
 
     public async updatePage(page: PageContract): Promise<void> {
+        if (!page) {
+            throw new Error(`Parameter "page" not specified.`);
+        }
+
         await this.objectStorage.updateObject<PageContract>(page.key, page);
     }
 
-    public async getPageContent(pageKey: string): Promise<Contract> {
-        const page = await this.getPageByKey(pageKey);
+    public async getPageContent(key: string): Promise<Contract> {
+        if (!key) {
+            throw new Error(`Parameter "key" not specified.`);
+        }
+
+        const page = await this.getPageByKey(key);
         return await this.objectStorage.getObject(page.contentKey);
     }
 
-    public async updatePageContent(pageKey: string, document: Contract): Promise<void> {
+    public async updatePageContent(pageKey: string, content: Contract): Promise<void> {
+        if (!pageKey) {
+            throw new Error(`Parameter "pageKey" not specified.`);
+        }
+
+        if (!content) {
+            throw new Error(`Parameter "content" not specified.`);
+        }
+
         const page = await this.getPageByKey(pageKey);
-        this.objectStorage.updateObject(page.contentKey, document);
+        this.objectStorage.updateObject(page.contentKey, content);
     }
 }
